Fixed hyparview ttl check

+++ src/main/java/crdts/utils/VectorClock.java
@@ -34,7 +34,7 @@ public class VectorClock {
     }
 
     public int getHostClock(Host host) {
-        return this.clock.getOrDefault(host,0);
+        return this.clock.getOrDefault(host,0).intValue();
     }
 
     public static MySerializer<VectorClock> serializer = new MySerializer<VectorClock>() {
@@ -44,7 +44,7 @@ public class VectorClock {
             out.writeInt(clocks.entrySet().size());
             for (Map.Entry<Host, Integer> entry : clocks.entrySet()) {
                 Host k = entry.getKey();
-                Integer v = entry.getValue();
+                int v = entry.getValue().intValue();
                 Host.serializer.serialize(k, out);
                 out.writeInt(v);
             }
